# backend-hookathon/openapi.yaml
openapi: 3.0.3
info:
  title: Aqua0 API
  description: |
    REST API for the Aqua0 cross-chain shared liquidity protocol built on 1inch Aqua.
    Provides endpoints for strategies, positions, swaps, metrics, tokens, rebalancing, and user management.
  version: 1.0.0
  contact:
    name: Aqua0
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.aqua0.xyz
    description: Production
...
security:
  - BasicAuth: []
    ApiKeyAuth: []
paths:
...
  /api/v1/positions/{wallet}:
    get:
      tags: [Positions]
      summary: Get positions by wallet address
      operationId: getPositions
      parameters:
        - $ref: "#/components/parameters/WalletPath"
        - $ref: "#/components/parameters/ChainFilter"
        - name: strategyHash
          in: query
          schema:
            $ref: "#/components/schemas/StrategyHash"
        - name: active
          in: query
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Positions and LP accounts
          content:
            application/json:
              schema:
                type: object
                required: [positions, lpAccounts]
                properties:
                  positions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Position"
                  lpAccounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/LPAccount"
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication for API access
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key required for /api/v1/* endpoints
  parameters:
    WalletPath:
      name: wallet
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/EthereumAddress"
...
    ChainFilter:
      name: chainId
      in: query
      schema:
        $ref: "#/components/schemas/NumericChainId"
...
  schemas:
...
    EthereumAddress:
      type: string
      pattern: "^0x[a-fA-F0-9]{40}$"
      example: "0x4200000000000000000000000000000000000006"
    StrategyHash:
      type: string
      pattern: "^0x[a-fA-F0-9]{64}$"
      example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
...
    NumericChainId:
      type: integer
      enum: [8453, 130]
...
    BigIntString:
      type: string
      pattern: "^\\d+$"
      description: Large integer represented as a string for JSON compatibility
      example: "1000000000000000000"
...
    Position:
      type: object
      required: [lpAccountAddress, strategyHash, token, chainId, balance]
      properties:
        lpAccountAddress:
          $ref: "#/components/schemas/EthereumAddress"
        owner:
          $ref: "#/components/schemas/EthereumAddress"
        strategyHash:
          $ref: "#/components/schemas/StrategyHash"
        token:
          $ref: "#/components/schemas/EthereumAddress"
        chainId:
          type: integer
        balance:
          $ref: "#/components/schemas/BigIntString"
        balanceUsd:
          type: number
          nullable: true
        depositedUsd:
          type: number
          nullable: true
        earnedFeesUsd:
          type: number
          nullable: true
        pnlUsd:
          type: number
          nullable: true
        pnlPercentage:
          type: number
          nullable: true
        lastSyncedAt:
          type: string
          nullable: true
    LPAccount:
      type: object
      required: [address, chainId, owner]
      properties:
        address:
          $ref: "#/components/schemas/EthereumAddress"
        chainId:
          type: integer
        owner:
          $ref: "#/components/schemas/EthereumAddress"
        rebalancerAuthorized:
          type: boolean
        rebalancerAddress:
          $ref: "#/components/schemas/EthereumAddress"
        createdAt:
          type: string
...